requirements.txt
shippo>=1.5.1
whaaaaat>=0.5.2
PyYAML>=3.12